Macros invocation from separate files (#22)
Macros loaded from libraries defined in proc_grps.json.

Diagnostics with "call stack" information

VS Code language contribution: we only parse files that are in HLASM language now.

+++ parser_library/src/checking/checker_helper.h
@@ -151,7 +151,7 @@ namespace checking
 		else
 			leap_year = false;
 
-		return (day <= days_in_month[month-1] || (month == 2 && leap_year && day == 29));
+		return ((size_t)day <= days_in_month[month-1] || (month == 2 && leap_year && day == 29));
 	};
 
 	static int get_size_of_operand(context::operand_format format)
